refactor: inline frontmatter parser into document.py, delete frontmatter.py
parse_frontmatter() and _parse_yaml_subset() now live in document.py —
the only module that uses them. Fewer files, cleaner ownership.
